backend/requirements.txt
Django==4.1.7
graphene-django==3.0.0
django-graphql-jwt==0.3.4
psycopg2-binary==2.9.5
django-cors-headers==3.14.0
requests==2.28.2
